# openapi.yaml
components:
  schemas:
    AcademicTitle:
      description: Academic title
      enum:
      - Dr.
      - Prof.
      title: AcademicTitle
      type: string
    AuthorityLevel:
      description: Type of asserting authority
      enum:
      - self
      - peer
      - system
      - so
      - dac
      title: AuthorityLevel
      type: string
    Claim:
      additionalProperties: false
      description: A claim about a user with a user ID and all data
      properties:
        asserted_by:
          anyOf:
          - $ref: '#/components/schemas/AuthorityLevel'
          - type: 'null'
          description: Authority level
          examples:
          - so
          - dac
          - system
        assertion_date:
          description: Date when the assertion was made
          examples:
          - '2022-11-30T12:00:00Z'
          format: date-time
          title: Assertion Date
          type: string
        conditions:
          anyOf:
          - items:
              items:
                $ref: '#/components/schemas/Condition-Output'
              type: array
            type: array
          - type: 'null'
          description: Set of conditions
          title: Conditions
        creation_date:
          description: Date of creation of this claim
          format: date-time
          title: Creation Date
          type: string
        id:
          description: Internal claim ID
          title: Id
          type: string
        iva_id:
          anyOf:
          - type: string
          - type: 'null'
          description: ID of an IVA associated with this claim
          title: Iva Id
        revocation_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: If revoked, date of revocation
          title: Revocation Date
        source:
          description: Asserting organization
          examples:
          - https://home.org
          format: uri
          maxLength: 2083
          minLength: 1
          title: Source
          type: string
        sub_source:
          anyOf:
          - type: string
          - type: 'null'
          description: Asserting sub-organization
          examples:
          - https://dac.home.org
          title: Sub Source
        user_id:
          description: Internal user ID
          title: User Id
          type: string
        valid_from:
          description: Start date of validity
          examples:
          - '2023-01-01T00:00:00Z'
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          examples:
          - '2023-12-31T23:59:59Z'
          format: date-time
          title: Valid Until
          type: string
        visa_type:
          $ref: '#/components/schemas/VisaType'
          examples:
          - AffiliationAndRole
        visa_value:
          anyOf:
          - type: string
          - type: 'null'
          description: Scope of the claim depending of the visa type
          examples:
          - faculty@home.org
          title: Visa Value
      required:
      - valid_from
      - valid_until
      - visa_type
      - visa_value
      - assertion_date
      - source
      - user_id
      - creation_date
      title: Claim
      type: object
...
    ClaimUpdate:
      additionalProperties: false
      description: A set of attributes that shall be updated in a claim.
      properties:
        revocation_date:
          description: Date of revocation
          format: date-time
          title: Revocation Date
          type: string
      required:
      - revocation_date
      title: ClaimUpdate
      type: object
    ClaimValidity:
      additionalProperties: false
      description: Start and end dates for validating claims.
      properties:
        valid_from:
          description: Start date of validity
          examples:
          - '2023-01-01T00:00:00Z'
          format: date-time
          title: Valid From
          type: string
        valid_until:
          description: End date of validity
          examples:
          - '2023-12-31T23:59:59Z'
          format: date-time
          title: Valid Until
          type: string
      required:
      - valid_from
      - valid_until
      title: ClaimValidity
      type: object
...
    Condition-Output:
      additionalProperties: false
      description: A single condition to check a type and a set of claims
      properties:
        matches:
          items:
            $ref: '#/components/schemas/ClaimMatch'
          title: Matches
          type: array
        type:
          $ref: '#/components/schemas/VisaType'
      required:
      - type
      - matches
      title: Condition
      type: object
...
    StatusChange:
      additionalProperties: false
      description: Details of a status change
      properties:
        by:
          anyOf:
          - type: string
          - type: 'null'
          description: ID of the user who changed the status
          title: Changed by
        change_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Date of last change
          title: Change Date
        context:
          anyOf:
          - type: string
          - type: 'null'
          description: Status change context
          title: Context
        previous:
          anyOf:
          - $ref: '#/components/schemas/UserStatus'
          - type: 'null'
          description: Previous user status
      title: StatusChange
      type: object
    User:
      additionalProperties: false
      description: Complete user model with ID
      properties:
        active_access_requests:
          default: []
          description: List of IDs of all active data access requests created by the
            user
          items:
            type: string
          title: Active Access Requests
          type: array
        active_submissions:
          default: []
          description: List of IDs of all active submissions created by the user
          items:
            type: string
          title: Active Submissions
          type: array
        email:
          description: Preferred e-mail address of the user
          examples:
          - user@home.org
          format: email
          title: Email
          type: string
        ext_id:
          description: External (Life Science) ID
          examples:
          - user@lifescience-ri.eu
          format: email
          title: External ID
          type: string
        id:
          description: Internal user ID
          title: Id
          type: string
        name:
          description: Full name of the user
          examples:
          - Rosalind Franklin
          title: Name
          type: string
        registration_date:
          description: Date when the user was registered
          format: date-time
          title: Registration Date
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
          description: Registration status of the user
        status_change:
          anyOf:
          - $ref: '#/components/schemas/StatusChange'
          - type: 'null'
          description: Last status change
        title:
          anyOf:
          - $ref: '#/components/schemas/AcademicTitle'
          - type: 'null'
          description: Academic title of the user
          title: Academic title
      required:
      - registration_date
      - name
      - email
      - ext_id
      - status
      title: User
      type: object
    UserBasicData:
      additionalProperties: false
      description: Basic data of a user
      properties:
        email:
          description: Preferred e-mail address of the user
          examples:
          - user@home.org
          format: email
          title: Email
          type: string
        name:
          description: Full name of the user
          examples:
          - Rosalind Franklin
          title: Name
          type: string
        title:
          anyOf:
          - $ref: '#/components/schemas/AcademicTitle'
          - type: 'null'
          description: Academic title of the user
          title: Academic title
      required:
      - name
      - email
      title: UserBasicData
      type: object
    UserModifiableData:
      additionalProperties: false
      description: User data that can be modified
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/UserStatus'
          - type: 'null'
          description: Registration status of the user
        title:
          anyOf:
          - $ref: '#/components/schemas/AcademicTitle'
          - type: 'null'
          description: Academic title of the user
          title: Academic title
      title: UserModifiableData
      type: object
...
    UserStatus:
      description: User status enum
      enum:
      - active
      - inactive
      title: UserStatus
      type: string
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  contact:
    email: helpdesk@ghga.de
    name: GHGA
  description: REST API for managing the GHGA users and user claims
  license:
    name: Apache 2.0
  title: Auth Service API
  version: 5.0.1
openapi: 3.1.0
paths:
  /download-access/users/{user_id}/datasets:
    get:
      description: Endpoint to get a list of the IDs of all datasets that can be downloaded
        by the given user. For internal use only.
      operationId: get_download_access_list
      parameters:
      - description: Internal ID of the user
        in: path
        name: user_id
        required: true
        schema:
          description: Internal ID of the user
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  type: string
                title: Response Get Download Access List
                type: array
          description: Dataset IDs with download access have been retrieved.
        '404':
          description: The user was not found or is inactive.
        '422':
          description: Validation error in submitted user IDs.
      summary: Get list of all dataset IDs with download access permission
      tags:
      - datasets
  /download-access/users/{user_id}/datasets/{dataset_id}:
    get:
      description: Endpoint to check whether the given dataset can be downloaded by
        the given user. For internal use only.
      operationId: check_download_access
      parameters:
      - description: Internal ID of the user
        in: path
        name: user_id
        required: true
        schema:
          description: Internal ID of the user
          title: User Id
          type: string
      - description: Internal ID of the dataset
        in: path
        name: dataset_id
        required: true
        schema:
          description: Internal ID of the dataset
          pattern: ^[A-Z]{1,6}[0-9]{3,18}$
          title: Dataset Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Check Download Access
                type: boolean
          description: Download access has been checked.
        '404':
          description: The user was not found or is inactive.
        '422':
          description: Validation error in submitted user IDs.
      summary: Check download access permission for a dataset
      tags:
      - datasets
  /download-access/users/{user_id}/ivas/{iva_id}/datasets/{dataset_id}:
    post:
      description: Endpoint to add a controlled access grant for a given dataset so
        that it can be downloaded by the given user with the given IVA. For internal
        use only.
      operationId: grant_download_access
      parameters:
      - description: The internal ID of the user
        in: path
        name: user_id
        required: true
        schema:
          description: The internal ID of the user
          title: User Id
          type: string
      - description: The ID of the IVA
        in: path
        name: iva_id
        required: true
        schema:
          description: The ID of the IVA
          title: Iva Id
          type: string
      - description: The ID of the dataset
        in: path
        name: dataset_id
        required: true
        schema:
          description: The ID of the dataset
          pattern: ^[A-Z]{1,6}[0-9]{3,18}$
          title: Dataset Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimValidity'
        required: true
      responses:
        '204':
          description: Download access has been granted.
        '404':
          description: The user or the IVA was not found.
        '422':
          description: Validation error in submitted user IDs.
      summary: Grant download access permission for a dataset
      tags:
      - datasets
...
  /users/{id}:
    delete:
      description: Endpoint used to delete a user. Can only be performed by a data
        steward.
      operationId: delete_user
      parameters:
      - description: Internal ID
        in: path
        name: id
        required: true
        schema:
          description: Internal ID
          title: Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '204':
          description: User data was successfully deleted.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized to delete this user.
        '404':
          description: The user was not found.
        '422':
          description: Validation error in submitted user identification.
      security:
      - HTTPBearer: []
      summary: Delete user
      tags:
      - users
    get:
      description: Endpoint used to get the user data for a specified user. Can only
        be performed by a data steward or the same user.
      operationId: get_user
      parameters:
      - description: Internal User ID
        in: path
        name: id
        required: true
        schema:
          description: Internal User ID
          title: Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Requested user has been found.
        '401':
          description: Not authorized to request user data.
        '403':
          description: Not authorized to request data of this user.
        '404':
          description: The user was not found.
        '422':
          description: Validation error in submitted user identification.
      security:
      - HTTPBearer: []
      summary: Get user data
      tags:
      - users
    patch:
      description: Endpoint used to modify the user data for a specified user. Can
        only be performed by a data steward or the same user.
      operationId: patch_user
      parameters:
      - description: Internal ID
        in: path
        name: id
        required: true
        schema:
          description: Internal ID
          title: Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserModifiableData'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '204':
          description: User data was successfully saved.
        '401':
          description: Not authenticated.
        '403':
          description: Not authorized to make this modification.
        '404':
          description: The user was not found.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Modify user data
      tags:
      - users
    put:
      description: Endpoint used to update a registered user when their name or email
        used by LS Login have changed. May only be performed by the users themselves.
        Data delivered by the OIDC provider may not be altered.
      operationId: put_user
      parameters:
      - description: Internal ID
        in: path
        name: id
        required: true
        schema:
          description: Internal ID
          title: Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserBasicData'
        required: true
      responses:
        '204':
          description: User was successfully updated.
        '401':
          description: Not authorized to update users.
        '403':
          description: Not authorized to update this user.
        '404':
          description: User does not exist.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Update a user
      tags:
      - users
  /users/{user_id}/claims:
    get:
      description: Endpoint used to get all claims for a specified user.
      operationId: get_claims
      parameters:
      - description: Internal ID of the user
        in: path
        name: user_id
        required: true
        schema:
          description: Internal ID of the user
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Claim'
                title: Response 200 Get Claims
                type: array
          description: User claims have been retrieved.
        '401':
          description: Not authorized to get user claims.
        '404':
          description: The user was not found.
        '422':
          description: Validation error in submitted user ID.
      summary: Get all claims for a given user
      tags:
      - claims
    post:
      description: Endpoint used to store a new claim about a user.
      operationId: post_claim
      parameters:
      - description: Internal ID of the user
        in: path
        name: user_id
        required: true
        schema:
          description: Internal ID of the user
          title: User Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreation'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Claim'
          description: Claim was successfully stored.
        '400':
          description: Claim cannot be stored.
        '404':
          description: The user was not found.
        '409':
          description: Claim was already stored.
        '422':
          description: Validation error in submitted ID or claims data.
      summary: Store a user claim
      tags:
      - claims
  /users/{user_id}/claims/{claim_id}:
    delete:
      description: Endpoint used to delete an existing user claim.
      operationId: delete_claim
      parameters:
      - description: Internal user ID
        in: path
        name: user_id
        required: true
        schema:
          description: Internal user ID
          title: User Id
          type: string
      - description: Internal claim ID
        in: path
        name: claim_id
        required: true
        schema:
          description: Internal claim ID
          title: Claim Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '204':
          description: User claim was successfully deleted.
        '404':
          description: The user claim was not found.
        '422':
          description: Validation error in submitted user or claim ID.
      summary: Delete an existing user claim
      tags:
      - claims
    patch:
      description: Endpoint used to revoke a claim for a specified user.
      operationId: patch_claim
      parameters:
      - description: Internal user ID
        in: path
        name: user_id
        required: true
        schema:
          description: Internal user ID
          title: User Id
          type: string
      - description: Internal claim ID
        in: path
        name: claim_id
        required: true
        schema:
          description: Internal claim ID
          title: Claim Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimUpdate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '204':
          description: User claim was successfully saved.
        '404':
          description: The user claim was not found.
        '422':
          description: Validation error in submitted user data.
      summary: Revoke an existing user claim
      tags:
      - claims
...